backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
google-generativeai==0.8.3
python-dotenv==1.0.1
python-dateutil==2.9.0
pydantic==2.6.4
